fix: preloading background texture and sizing monkey sprite loading spinner
