+++ src/components/overlay/index.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { useRef, useEffect } from "react";
+import { useRef, useEffect, useState } from "react";
 import textBox from "./assets/text-box.png";
 import Arrows from "@/components/overlay/Arrows";
 import Image from "next/image";
@@ -7,9 +7,11 @@ import CharacterEvents from "@/components/overlay/CharacterEvents";
 import Milestones from "@/components/overlay/Milestone";
 import Socials from "@/components/overlay/Socials";
 import BottomTextBox from "./BottomTextBox";
+import PreloadTextures from "../utils/PreloadTextures";
 
 export default function UIOverlay() {
   const bgAudioRef = useRef<HTMLAudioElement>(null);
+  const [texturesReady, setTexturesReady] = useState(false);
 
   useEffect(() => {
     if (bgAudioRef.current) {
@@ -18,6 +20,17 @@ export default function UIOverlay() {
       });
     }
   }, []);
+
+  if (!texturesReady) {
+    return (
+      <>
+        <PreloadTextures onReady={setTexturesReady} />
+        <div style={{ textAlign: 'center', marginTop: '50px' }}>
+          <h1>Loading...</h1>
+        </div>
+      </>
+    );
+  }
 
   return (
     <div className="absolute top-0 left-0 w-full h-full pointer-events-none">
